requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# API Documentation
drf-spectacular==0.27.1

# Security & CORS
django-cors-headers==4.3.1

# Environment variables
python-decouple==3.8

# Utilities
PyJWT==2.8.0
cryptography==42.0.0

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Development
django-extensions==3.2.3
